requirements.txt
attrs==23.2.0
aws-cdk-lib==2.130.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-lambda-typing==2.18.0
boto3==1.34.50
botocore==1.34.50
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constructs==10.3.0
cryptography==35.0.0
et-xmlfile==1.1.0
idna==3.6
importlib_resources==6.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsii==1.94.0
MarkupSafe==2.1.5
openpyxl==3.1.2
publication==0.0.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==2.0.7
